requirements.txt
Django==1.8.3
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
Pillow==2.5.0
django-allauth
